feat: initial lubrication.comp file implementation
The comp file wires everything up, which is done without any testing yet.

+++ Components/src/Lubrication/lubrication_logic.h
@@ -34,9 +34,9 @@ typedef struct {
 /* The output signals from the lubrication pump logic */
 typedef struct {
     LubricationStates state; /* The current lubrication state. */
-    float buildingPressureStartTime;
-    float lubricationStartTime;
-    float lastCycleEndTime;
+    double buildingPressureStartTime;
+    double lubricationStartTime;
+    double lastCycleEndTime;
 } LubricationState;
 
 void lubricate(
